+++ R/matching_df.R
@@ -5,8 +5,9 @@
 #' @return A list of "matching" class containing
 #'     (1) a data frame of the matching results,
 #'     (2) a character string showing which algorithm was used,
-#'     (3) a character string of the matching results, and
-#'     (4) a character string of the history of matching steps.
+#'     (3) a character string of the matching results,
+#'     (4) a character string of the history of matching steps, and
+#'     (5) a list of preferences of each group.
 #' @param df1 A data frame or a data file containing preferences of
 #'     the proposers.
 #' @param df2 A data frame or a data file containing preferences of